requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pyjwt
slowapi
bcrypt
